requirements.txt
coloredlogs==15.0.1
flatbuffers==25.2.10
humanfriendly==10.0
mpmath==1.3.0
numpy==1.26.4
onnxruntime-gpu==1.16.1
packaging==25.0
protobuf==6.31.1
pyreadline3==3.5.4
sympy==1.14.0
opencv-python==4.10.0.84
Pillow==10.4.0
insightface==0.7.3
gfpgan==1.3.8
torch==2.4.0
